+++ js/packages/core/src/ufg/get-base-eyes.ts
@@ -1,36 +1,49 @@
-import type {OpenSettings} from './types'
-import type {Core as CoreBase, Eyes as EyesBase} from '@applitools/core-base'
+import type {Eyes, OpenSettings} from './types'
+import type {Eyes as BaseEyes} from '@applitools/core-base'
+import {type SpecType} from '@applitools/driver'
 import {type Logger} from '@applitools/logger'
-import {type UFGClient, type Renderer} from '@applitools/ufg-client'
+import {type RendererSettings} from '@applitools/ufg-client'
 import * as utils from '@applitools/utils'
 
-type Options = {
+type Options<TSpec extends SpecType> = {
   settings: OpenSettings
-  eyes?: EyesBase[]
-  core: CoreBase
-  client: UFGClient
+  eyes: Eyes<TSpec>
+  base?: BaseEyes[]
   logger: Logger
 }
 
-export function makeGetBaseEyes({settings: defaultSettings, core, client, eyes, logger: defaultLogger}: Options) {
+export function makeGetBaseEyes<TSpec extends SpecType>({
+  settings: defaultSettings,
+  eyes,
+  base,
+  logger: defaultLogger,
+}: Options<TSpec>) {
   const getBaseEyesWithCache = utils.general.cachify(getBaseEyes, ([options]) => options?.settings)
-  if (eyes) eyes.forEach(eyes => getBaseEyesWithCache.setCachedValue(eyes.test.rendererInfo, Promise.resolve([eyes])))
+  if (base) {
+    base.forEach(baseEyes =>
+      getBaseEyesWithCache.setCachedValue(baseEyes.test.rendererInfo, Promise.resolve([baseEyes])),
+    )
+  }
   return getBaseEyesWithCache
 
   async function getBaseEyes({
     settings,
     logger = defaultLogger,
   }: {
-    settings?: {type: 'web' | 'native'; renderer: Renderer}
+    settings?: RendererSettings
     logger?: Logger
-  } = {}): Promise<EyesBase[]> {
+  } = {}): Promise<BaseEyes[]> {
     logger.log(`Command "getBaseEyes" is called with settings`, settings)
     if (!settings) throw new Error('')
-    const environment = await client.bookRenderer({settings})
-    const eyes = await core.openEyes({
+    const ufgClient = await eyes.core.getUFGClient({
+      config: {...eyes.test.account, ...eyes.test.account.ufg, proxy: eyes.test.server.proxy},
+      logger,
+    })
+    const environment = await ufgClient.bookRenderer({settings})
+    const baseEyes = await eyes.core.base.openEyes({
       settings: {...defaultSettings, environment: {...defaultSettings.environment, ...environment}},
       logger,
     })
-    return [eyes]
+    return [baseEyes]
   }
 }