+++ js/packages/core/src/ufg/open-eyes.ts
@@ -1,8 +1,8 @@
-import type {DriverTarget, Eyes, OpenSettings, TestInfo} from './types'
-import type {Core as BaseCore, Eyes as BaseEyes} from '@applitools/core-base'
+import type {DriverTarget, Core, Eyes, OpenSettings, TestInfo} from './types'
+import type {Eyes as BaseEyes} from '@applitools/core-base'
 import {type Logger} from '@applitools/logger'
+import {type Renderer} from '@applitools/ufg-client'
 import {makeDriver, type SpecType, type SpecDriver} from '@applitools/driver'
-import {makeUFGClient, Renderer, type UFGClient} from '@applitools/ufg-client'
 import {makeGetBaseEyes} from './get-base-eyes'
 import {makeCheck} from './check'
 import {makeCheckAndClose} from './check-and-close'
@@ -13,31 +13,30 @@ import {AbortController} from 'abort-controller'
 import * as utils from '@applitools/utils'
 
 type Options<TSpec extends SpecType> = {
-  core: BaseCore
-  client?: UFGClient
+  core: Core<TSpec>
   spec?: SpecDriver<TSpec>
   logger: Logger
 }
 
-export function makeOpenEyes<TSpec extends SpecType>({core, client, spec, logger: defaultLogger}: Options<TSpec>) {
+export function makeOpenEyes<TSpec extends SpecType>({core, spec, logger: defaultLogger}: Options<TSpec>) {
   return async function openEyes({
     target,
     settings,
-    eyes,
+    base,
     logger = defaultLogger,
   }: {
     target?: DriverTarget<TSpec>
     settings: OpenSettings
-    eyes?: BaseEyes[]
+    base?: BaseEyes[]
     logger?: Logger
   }): Promise<Eyes<TSpec>> {
     logger.log(
       `Command "openEyes" is called with ${target ? 'default driver and' : ''}`,
       ...(settings ? ['settings', settings] : []),
-      eyes ? 'predefined eyes' : '',
+      base ? 'predefined eyes' : '',
     )
     const driver = target && (await makeDriver({spec, driver: target, logger}))
-    if (driver && !eyes) {
+    if (driver && !base) {
       const currentContext = driver.currentContext
       settings.environment ??= {}
       if (driver.isEC) {
@@ -50,34 +49,26 @@ export function makeOpenEyes<TSpec extends SpecType>({core, client, spec, logger
     }
     const controller = new AbortController()
     const account = await core.getAccountInfo({settings, logger})
-    client ??= makeUFGClient({
-      config: {...account.ufg, ...account, proxy: settings.proxy},
-      concurrency: settings.renderConcurrency ?? 5,
-      logger,
-    })
-
-    const getBaseEyes = makeGetBaseEyes({settings, eyes, core, client, logger})
     return utils.general.extend({}, eyes => {
       const storage = new Map<string, Promise<{renderer: Renderer; eyes: BaseEyes}>[]>()
-
       let running = true
-
       return {
         type: 'ufg' as const,
-        test: <TestInfo>{
+        core,
+        test: {
           userTestId: settings.userTestId,
           batchId: settings.batch?.id,
           keepBatchOpen: settings.keepBatchOpen,
           server: {serverUrl: settings.serverUrl, apiKey: settings.apiKey, proxy: settings.proxy},
           account,
-        },
+        } as TestInfo,
         get running() {
           return running
         },
-        getBaseEyes,
+        getBaseEyes: makeGetBaseEyes({settings, eyes, base, logger}),
         // check with indexing and storage
         check: utils.general.wrap(
-          makeCheck({eyes, client: client!, target: driver, spec, signal: controller.signal, logger}),
+          makeCheck({eyes, target: driver, spec, signal: controller.signal, logger}),
           async (check, options = {}) => {
             const results = await check(options)
             results.forEach(result => {
@@ -88,14 +79,7 @@ export function makeOpenEyes<TSpec extends SpecType>({core, client, spec, logger
           },
         ),
         checkAndClose: utils.general.wrap(
-          makeCheckAndClose({
-            eyes,
-            client: client!,
-            target: driver,
-            spec,
-            signal: controller.signal,
-            logger,
-          }),
+          makeCheckAndClose({eyes, target: driver, spec, signal: controller.signal, logger}),
           async (checkAndClose, options = {}) => {
             const results = await checkAndClose(options)
             results.forEach(result => {
@@ -106,12 +90,14 @@ export function makeOpenEyes<TSpec extends SpecType>({core, client, spec, logger
           },
         ),
         close: utils.general.wrap(makeClose({storage, target: driver, logger}), async (close, options) => {
+          if (!running) return
           running = false
           return close(options)
         }),
         abort: utils.general.wrap(
           makeAbort({storage, target: driver, spec, controller, logger}),
           async (abort, options) => {
+            if (!running) return
             running = false
             return abort(options)
           },