+++ js/packages/ufg-client/src/types.ts
@@ -264,6 +264,8 @@ export type RenderSettings = RendererSettings & {
   rendererId: string
   region?: Region | Selector
   fully?: boolean
+  scrollRootElement?: Selector
+  stitchMode?: 'Scroll' | 'CSS' | 'Resize'
   selectorsToCalculate?: Selector[]
   includeFullPageSize?: boolean
   ufgOptions?: Record<string, any>