+++ python/eyes_images/applitools/selenium/schema.py
@@ -55,13 +55,6 @@ from .schema_fields import (
 
 if t.TYPE_CHECKING:
     from applitools.common import config
-    from applitools.common.utils.custom_types import ViewPort
-    from applitools.core import locators
-    from applitools.images import fluent as ics
-    from applitools.selenium.fluent import selenium_check_settings as cs
-    from applitools.selenium.optional_deps import WebDriver
-
-    from ..core import batch_close
 
 
 # Default marshmallow.Schema has no option to skip attributes with None value
@@ -372,6 +365,7 @@ class CheckSettings(USDKSchema):
     accessibility_regions = List(
         Nested(AccessibilityRegionReference), dump_to="accessibilityRegions"
     )
+    webview = Field()
 
 
 class LocateSettings(USDKSchema):